client/__tests__/types/api/score.test.ts
import { getDanceLevel } from '~/types/api/score'

describe('./types/api/score.ts', () => {
  describe('getDanceLevel', () => {
    test.each([
      [0, 'D'],
      [550000, 'D+'],
      [590000, 'C-'],
      [600000, 'C'],
      [650000, 'C+'],
      [690000, 'B-'],
      [700000, 'B'],
      [750000, 'B+'],
      [790000, 'A-'],
      [800000, 'A'],
      [850000, 'A+'],
      [890000, 'AA-'],
      [900000, 'AA'],
      [950000, 'AA+'],
      [990000, 'AAA'],
      [1000000, 'AAA'],
    ])('(%i) returns "%s"', (score, expected) => {
      expect(getDanceLevel(score)).toBe(expected)
    })
  })
})
